# Reimplement DiscreteDistribution.__imul__