Avoiding unnecessary reinitialization decreases the time needed to make predictions by about 50%.

Test program:

```
import numpy as np
from neurtu import delayed, Benchmark
from pomegranate import BayesianNetwork
from random import random
from sklearn.datasets import load_digits

X = load_digits(return_X_y=True)[0][:,:10]
net = BayesianNetwork().from_samples(X)

predict = delayed(net.predict_proba)({str(x): 0 for x in range(5)})

print(Benchmark(wall_time=True, cpu_time=True, repeat=50)(predict))
```

Before:

```
      wall_time  cpu_time                                                                                                     
mean   0.002891  0.002878
max    0.002938  0.002927
std    0.000013  0.000011
```

After:

```
      wall_time  cpu_time                                                                                                     
mean   0.001496  0.001493
max    0.001551  0.001557
std    0.000011  0.000014
```

I took care to keep the behavior identical to the original, but even more time could be saved if the summaries etc. do not need to be reset.

I also threw in a small cleanup that simplifies getting the data type name in the DiscreteDistribution constructor and a couple of other places. Let me know if you'd rather review that in a separate pull request though.