@alexhenrie 2020-04-22
Okay, after fixing the bugs there doesn't appear to actually be any noticeable speed improvement to reimplementing `__imul__`, so I am closing this pull request and will instead open a new one with just the type name cleanup.